Fix server deserializer
The server deserializer has the same behavior as the client deserializer
but also rejects unknown fields, so we previously tried to have the
server deserializer wrap the client deserializer. However, we
accidentally added an extra layer of nesting at one of the wrapping
stages so you'd end up with an exponential number of nested visitors as
the type heirarchy increasees.

We could have fixed that, but it's really hard to think about where the
nesting needs to be applied, and easier to just implement them both
separately side by side.