Service client generation
We want to enable code generation to evolve separately from the HTTP
client/server logic, so the conjure-http crate just contains a tiny shim
interface that we generate code targetting and the real clients/servers
work with generically.

In the case of clients, we generate a struct which wraps a generic
`Client` implementation and has methods per endpoint that handle the
serialization and deserialization logic.