Expose conjure tags as endpoint parameters
Tags were previously not exposed at the endpoint level. The only tags
that were effective were 'server-request-context' and
'server-limit-request-size', both of which were interpreted by
conjure-rust at codegen time.

Add a 'tags' function to Endpoint, which exposes the set of tags. This
is similar to the java undertow generator.

Forgoes adding 'tags' to the client, unlike the java dialogue
implementation. This is because the clients are used directly in rust,
as opposed to being passed into a dialogue client generator.